pyproject.toml
[tool.poetry]
name = "dsip-itp"
version = "0.1.0"
description = ""
authors = ["Lorenz191 <144610453+Lorenz191@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
pre-commit = "^3.8.0"
djangorestframework = "^3.15.2"
torch = "^2.4.1"
transformers = "^4.45.1"
german = "^0.1.0"
sentiment = "^0.0.1"
germansentiment = "^1.1.0"
pandas = "^2.2.3"
numpy = "^2.1.1"
matplotlib = "^3.9.2"
mysql-connector-python = "^9.1.0"
python-dotenv = "^1.0.1"
pillow = "^11.0.0"
mysqlclient = "^2.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
